requirements.txt
beautifulsoup4==4.6.0
click==6.7
ezodf==0.3.2
Flask==0.12.2
Flask-Stache==0.1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
libsass==0.14.5
lxml==4.7.1
MarkupSafe==1.1.0
numpy==1.22.2
odfpy==1.3.6
pandas==1.4.1
pandas-ods-reader==0.1.4
pystache==0.5.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
